+++ paver/tests/test_tasks.py
@@ -302,6 +302,21 @@ def test_consume_args():
     env = _set_environment(t3=t3)
     tasks._process_commands("t3 -v 1".split())
     assert t3.called
+
+def test_optional_args_in_tasks():
+    @tasks.task
+    def t1(options, optarg=None):
+        assert optarg is None
+
+    @tasks.task
+    def t2(options, optarg1='foo', optarg2='bar'):
+        assert optarg1 is 'foo'
+        assert optarg2 is 'bar'
+
+    env = _set_environment(t1=t1, t2=t2)
+    tasks._process_commands(['t1', 't2'])
+    assert t1.called
+    assert t2.called
     
 def test_debug_logging():
     @tasks.task